feat: handle signin logic in the same place
